Renamed EditPageSetting to better suit usage
